refactor: move server to a separate file
